requirements.txt
Flask==1.1.1
atomicwrites==1.3.0
attrs==19.3.0
blinker==1.4
captcha==0.3
Click==7.0
colorama==0.4.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-WTF==0.14.2
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
mysql==0.0.2
mysqlclient==1.4.4
mysql-connector==2.2.9
packaging==19.2
Pillow==6.2.1
pip==19.0.3
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.5
pytest==5.2.3
setuptools==40.8.0
six==1.13.0
SQLAlchemy==1.3.10
utils==0.9.0
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0